requirements.txt
aiohttp~=3.10.11
base58~=2.1.1
python-dotenv~=1.0.1
solana~=0.35.0
solders~=0.21.0
Telethon~=1.37.0
pandas~=2.2.3
scikit-learn~=1.5.2
pyyaml~=6.0.2
websockets~=10.4
numpy~=2.0.2
ydata-profiling~=4.12.0
redis~=5.2.0
rq~=2.0.0
python-binance
tabulate~=0.9.0
dune-client
python-logstash-async~=3.0.0
psycopg2